Maybe last commit
Honestly an SRPG is quite an undertaking; that and I'm not really
convinced that I'm up to the task in designing it. For now, if there's
one thing is that lager is quite an amazing library that I learnt a lot
from (holy shit meta programming batman)

It works, you need some content; I may update that, in the meantime I'll
just leave this here if you're interested into what I'm up to.